// src/app/core/models/agenda-row.model.ts
export interface AgendaRow {
  dia1: string;
  semana1: string;
  mes1: string;
  frase1: string;
  dia2: string;
  semana2: string;
  mes2: string;
  frase2: string;
  dia1v: string;
  semana1v: string;
  mes1v: string;
  frase1v: string;
  dia2v: string;
  semana2v: string;
  mes2v: string;
  frase2v: string;
}